Indicate goldfish drivers depend on HAS_IOMEM
Goldfish drivers make use of iomem to communicate with the host.
Indicate this dependency in the Kconfig files so that drivers are not
built on platforms that do not support this. This is useful for usermode
Linux which is used to run test cases on the kernel. These drivers do
not compile in usermode without this fix because iomem is not available.

Bug: 130662334
Bug: 116799743
Bug: 120912403
Change-Id: Ic785d050a7754942ec3250690c625b44d5dae335
Signed-off-by: Bjoern Johansson <bjoernj@google.com>
Signed-off-by: Roman Kiryanov <rkir@google.com>