Indicate goldfish drivers depend on HAS_IOMEM
Goldfish drivers make use of iomem to communicate with the host.
Indicate this dependency in the Kconfig files so that drivers are not
built on platforms that do not support this. This is useful for usermode
Linux which is used to run test cases on the kernel. These drivers do
not compile in usermode without this fix because iomem is not available.

Bug: 111742629
Change-Id: Ia66995e7bc718742f5f0cc6bf39a1641cba1b4f1
Signed-off-by: Bjoern Johansson <bjoernj@google.com>
Signed-off-by: Roman Kiryanov <rkir@google.com>