Indicate goldfish drivers depend on HAS_IOMEM
Goldfish drivers make use of iomem to communicate with the host.
Indicate this dependency in the Kconfig files so that drivers are not
built on platforms that do not support this. This is useful for usermode
Linux which is used to run test cases on the kernel. These drivers do
not compile in usermode without this fix because iomem is not available.

Bug: 116799743
Bug: 120912403
Change-Id: Ie83bb651764862157dd56b265018f2a314de3e81
Signed-off-by: Bjoern Johansson <bjoernj@google.com>
Signed-off-by: Roman Kiryanov <rkir@google.com>