Revert "xfrm: Do not parse 32bits compiled xfrm netlink msg on 64bits host"
With the compatibility code in netd this change is no longer needed and
will instead block the correctly aligned messages. This change is not
portable anywhere but will serve to make the emulator kernel work
together with the framework so we can support IPSec on x86_32 on a
x86_64 kernel.

This reverts commit 74005991b78a0a7a6546004fb37d33a651c613e7

Bug: 111742629
Change-Id: Ieabce432d8377b66ae7e6f3a8c0f5a29d398ec5d
Signed-off-by: Roman Kiryanov <rkir@google.com>