Revert "xfrm: Do not parse 32bits compiled xfrm netlink msg on 64bits host"
With the compatibility code in netd this change is no longer needed and
will instead block the correctly aligned messages. This change is not
portable anywhere but will serve to make the emulator kernel work
together with the framework so we can support IPSec on x86_32 on a
x86_64 kernel.

This reverts commit 74005991b78a0a7a6546004fb37d33a651c613e7

Bug: 130662334
Bug: 116799743
Bug: 120912403
Change-Id: I9d8276a25ae2d76c3c913b4691e570f41058a5c3
Signed-off-by: Roman Kiryanov <rkir@google.com>